+++ src/raw_cbc.c
@@ -43,10 +43,10 @@ typedef struct {
 } CbcModeState;
 
 EXPORT_SYM int CBC_start_operation(BlockBase *cipher,
-                    const uint8_t iv[],
-                    size_t iv_len,
-                    CbcModeState **pResult) {
-
+                                   const uint8_t iv[],
+                                   size_t iv_len,
+                                   CbcModeState **pResult)
+{
     if ((NULL == cipher) || (NULL == iv) || (NULL == pResult)) {
         return ERR_NULL;
     }
@@ -67,10 +67,10 @@ EXPORT_SYM int CBC_start_operation(BlockBase *cipher,
 }
 
 EXPORT_SYM int CBC_encrypt(CbcModeState *cbcState,
-            const uint8_t *in,
-            uint8_t *out,
-            size_t data_len) {
-
+                           const uint8_t *in,
+                           uint8_t *out,
+                           size_t data_len)
+{
     uint8_t *pt, *iv;
     size_t block_len;
 
@@ -108,10 +108,10 @@ EXPORT_SYM int CBC_encrypt(CbcModeState *cbcState,
 }
 
 EXPORT_SYM int CBC_decrypt(CbcModeState *cbcState,
-            const uint8_t *in,
-            uint8_t *out,
-            size_t data_len) {
-
+                           const uint8_t *in,
+                           uint8_t *out,
+                           size_t data_len)
+{
     uint8_t *pt, *iv;
     size_t block_len;
 
@@ -151,6 +151,9 @@ EXPORT_SYM int CBC_decrypt(CbcModeState *cbcState,
 
 EXPORT_SYM int CBC_stop_operation(CbcModeState *state)
 {
+    if (NULL == state)
+        return ERR_NULL;
+
     state->cipher->destructor(state->cipher);
     free(state);
     return 0;